fix style according to
https://codereview.stackexchange.com/questions/242742/changing-values-in-a-vector/242743#242743

+++ quickfind/tests/lib_test.rs
@@ -3,8 +3,8 @@ mod tests {
     use quickfind::*;
 
     struct ConnectedRecord {
-        left: u64,
-        right: u64,
+        left: usize,
+        right: usize,
         is_connected: bool,
     }
